refactor(scripts): extract stableStringify into shared scripts/lib module
Sibling-pattern check on the prior commit found scripts/lib/job-identity.mjs
already had its own copy of the same recursive key-sort stringify helper I'd
just added inline to cf-locale-failover-setup.mjs. Both now import from a new
scripts/lib/stable-stringify.mjs. scripts/lib/git-commit-data.sh keeps its own
CommonJS twin (documented, deliberately diverges on undefined-handling for its
merge-fingerprint use case) since it runs via `node -` on stdin and can't
import an ESM module — same twin pattern this repo already uses for
BOT_GATE_FN / isLikelyBot().
